feat(env): float source placement + Euclidean/continuous-field sensing (T6 group 2)
Lifts the three continuous-2D deferrals from T5 (logbook 027 / spec):
- Float source placement: Continuous2DEnvironment overrides _generate_food_candidate
  to place sources at real-valued coords (uniform float in [0, grid_size-1]),
  reusing the inherited Euclidean validity/Poisson machinery. Grid base keeps its
  integer source type.
- True Euclidean: get_nearest_food_distance_for drops the round() (returns float).
- Continuous-field sampling: _continuous_lateral_offsets now samples real-valued
  points from the worm's continuous position (no integer-cell snap) against the
  continuous field; the analytic field getters (food/predator/temperature/oxygen/
  pheromone concentration + gradients, in env/temperature/oxygen/pheromone) widen
  their position param to the float Position alias (int is assignable to float, so
  grid callers are unaffected).

self.foods stays integer-annotated (the grid contract); the continuous subclass
stores floats at runtime — coordinate-compatible everywhere (every consumer does
Euclidean/arithmetic). The two cell-indexing sites (grid render goals) snap to a
cell first then bounds-check, so a near-edge float can't index out of range.

Tests: float placement (in-bounds, real-valued, candidate generator), un-rounded
Euclidean distance, float continuous-field lateral sampling. Full agent+env+brain
sweep green (2208 passed). Implements tasks 2.1-2.3.

+++ packages/quantum-nematode/quantumnematode/env/pheromone.py
@@ -121,7 +121,7 @@ class PheromoneField:
 
     def get_concentration(
         self,
-        position: tuple[int, int],
+        position: tuple[float, float],
         current_step: int,
     ) -> float:
         """Compute pheromone concentration at a position.
@@ -159,7 +159,7 @@ class PheromoneField:
 
     def get_gradient(
         self,
-        position: tuple[int, int],
+        position: tuple[float, float],
         current_step: int,
     ) -> tuple[float, float]:
         """Compute pheromone gradient via central differences.
@@ -185,7 +185,7 @@ class PheromoneField:
 
     def get_gradient_polar(
         self,
-        position: tuple[int, int],
+        position: tuple[float, float],
         current_step: int,
     ) -> tuple[float, float]:
         """Compute pheromone gradient in polar coordinates.
